LoRA-Recycle/requirements.txt
torch
torchvision
torchaudio
numpy
pandas
tqdm
scikit-learn
matplotlib
seaborn
opencv-python
pillow
timm
transformers
safetensors
termcolor
